Refactor CLI directory structure
