+++ test/cli/run.ts
@@ -3,7 +3,7 @@ import * as path from "path";
 
 jest.setTimeout(20000);
 
-const cliPath = path.join(__dirname, "../../src/tstl.ts");
+const cliPath = path.join(__dirname, "../../src/cli/tstl.ts");
 
 const defaultArgs = ["--skipLibCheck", "--types", "node"];
 export function forkCli(args: string[]): ChildProcess {
